requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-groq
langchain-google-genai
faiss-cpu
pypdf
sentence-transformers
python-dotenv
google-generativeai
